fixed color certificate compute by changing the state index
moving from std to cista changes iteration order. so the relevant state in the previous test at index 49 is now at index 75. In general, the abstraction states as a list do not align when comparing the std-based and cista-based versions of the library. this should be observed in the future.